Create maze code
Micro just bought a maze, that can be represented as a matrix A of size n x m , where rows are numbered from 1 to N and columns are numbered from 1 to M. Each cell of the matrix can be either 0 or 1. If a cell is 0, that means it cannot be visited and if it is 1, then it can be visited. Allowed moves are up, down, left and right. Help Micro to find out if he can reach from the cell(1,1)  to the cell(n,m) , it is guaranteed that the cells (1,1) and (n x m)have value 1.